Program_QSPI.bat: Add color and auto detect JLink path
I thought some color would be nice, but I had to create a DOS app
to change the color on the command line.
Also, since Segger is constantly updating the JLink installer, I switched
to using the registry to find out where it is installed instead of making
users edit the bat file.
Lastly, the batch file now loops back to the beginning in case people are
programming multiple images at once.