Add mem application
This application uses /dev/mem in order to allow users access to register
and memory locations from user space. It has been included in all the BSP
releases in binary form. However, the source is needed for rebuilding in
case you are using toolchain/glibc other than the BSP default.