build.sh: Add more programming lines for the jlink option
These are the lines that you copy/paste into u-boot after you download
to SDRAM so you can program into QSPI.

Show the programming lines for both a 5MB and 6MB kernel image sizes.

Show the programming lines for rootfs images all the way up to 32MB which
is the max because there is only 32MB of SDRAM on the RSK to download
to.