build.sh: use git to download uboot and kernel
Instead of downloading the official released versions of u-boot and kernel
and then applying our BSP patches, just get everything from our github
repositories.
This also makes it easier for people to pull in the latest updates
in between BSP releases.
While the entire repositories will be downloaded, git will only check out a
specific tested version instead of the current HEAD which might not be as
stable.