build.sh: Add LOCALVERSION= to kernel build command line
We do this to get rid of the + sign at the end of the kernel version string
that will appear if any source file is changed since now the kernel
resides under a git repo. The kernel build system wants to mark that the
kernel was built from code that was not checked into the repo....but most
users won't care about that.